+++ Tradgardsgolf.PipelineMessenger/Handlers/IHandler.cs
@@ -4,6 +4,6 @@ namespace Tradgardsgolf.PipelineMessenger.Handlers;
 
 public interface IHandler
 {
-    int Score(IMessage message, HandlerResult previousResult);
+    double Score(IMessage message, HandlerResult previousResult);
     HandlerResult Handle(IMessage message, HandlerResult previousResult);
 }
